split TransactionsPageTest on 2 parts
